new commit

+++ src/main/java/com/example/shopping_app_backend/entity/Item.java
@@ -22,7 +22,7 @@ public class Item {
     private String shop;
     private String productName;
     private BigDecimal price;
-    private int amount;
+    private BigDecimal amount;
     private String description;
     private String imageUrl;
     @ManyToOne
